requirements.txt
torch
torchvision
matplotlib
numpy
opencv-python # cv2
imageio
imageio-ffmpeg
pyyaml # yaml
tqdm
argparse
cython
scikit-image # skimage
scipy
onnxruntime
gradio

# LiteRT dependencies (for TDDFA_LiteRT)
# Install with: pip install ai-edge-torch-nightly ai-edge-litert
ai-edge-torch-nightly
ai-edge-litert
